Edit: CredentialsStorage.h and Storage.h now aren't pure virtual and they have example implementation with Btree as storage
Remove: UserRepository.h and ExampleStorage.h that implements the pure virtual classes previously
